requirements.txt
absl-py==0.6.1
astor==0.7.1
certifi==2018.10.15
chardet==3.0.4
configparser==3.5.0
cycler==0.10.0
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
idna==2.7
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.0
mysql-connector-python==8.0.12
numpy==1.15.4
opencv-python==3.4.3.18
Pillow==5.3.0
protobuf==3.6.1
psutil==5.4.8
pycairo==1.18.0
PyGObject==3.30.2
pyparsing==2.3.0
python-dateutil==2.7.3
python-debian==0.1.32
requests==2.20.0
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
